+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 Changelog
 =========
 
+Version 0.1.1.0
+---------------
+
+*Mar 25, 2018*
+
+<https://github.com/mstksg/hmatrix-backprop/releases/tag/v0.1.1.0>
+
+*   `dvmap` and `dmmap` rewritten to actually be usable, since before they had
+    nonsensical constraints.
+*   *backprop* types re-exported for convenience
+
+*Internal*
+
+*   Rewrote much code to use `isoVar` instead of `opIso`, for increased clarity
+    and cleanliness.  Requires *backprop-0.1.4.0*.
+
 Version 0.1.0.0
 ---------------